add user authorization

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,12 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing) {
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
+
         $formFields = $request->validate([
             'company' => 'required',
             'title' => 'required',
@@ -78,6 +84,11 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!')
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
